Extract shared draft composer builder
Collapse the ~45-line new-comment composer duplicated in both margins into
ui/draft-composer.ts; each margin supplies only its submit/cancel behavior.
